Extract useChatFileSources from FileUploadWithTokenCheck
